Sync to preview2-30335, mapping changes
Some work on type mapping to make everything work.
Every mapping now has to be one-to-one: a mapping returned for a CLR
type must contain that same type; previously when looking up by
DateTimeOffset we would return a timestamptz mapping which specified
CLR DateTime.

We now have separate mappings for DateTimeOffset and DateTime
(as well as for byte and short), and can handle multiple mappings
given a store type.